Fix operator precedence errors (#4)
a or b and c should be parsed as a or (b and c)

Although it's not specified at all in the PEP https://peps.python.org/pep-0508/#environment-markers
We can infer from the examples that: and has higher precedence than or, otherwise it's left-associative.

This fixes the logic to match. For the underlying packaging.markers implementation, it doesn't handle this directly when parsing. Instead, if you have a or b and c, you end up with an array of [a, or, b, and, c] so we need to implement the logic properly ourselves.

The easiest solution is to make a two pass through the array, replacing a and b with (a and b) first, and then a or b with (a or b) second to enforce evaluation order